pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "synthcal"
version = "0.1.0"
description = "Synthetic dataset generator for camera + laser-stripe calibration."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "synthcal contributors" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "numpy>=1.23",
  "PyYAML>=6.0",
  "Pillow>=10.0",
  "matplotlib>=3.7",
]

[project.scripts]
synthcal = "synthcal.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5",
  "build>=1.0",
  "twine>=5.0",
]
scipy = [
  "scipy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/synthcal"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["synthcal"]
